Add drag mode, enable mouse actions in helper window.
In the new drag mode, when user drags on the screen, instead of
selecting text, we drag the document content along with the mouse
cursor (the default keybinding to switch the drag mode on and off
is <f6>).
Also we add the ability to quickly modify portals directly in the
window by using mouse wheel or dragging the mouse cursor. (zooming
using control+mousewheel is also supported).