Split the database into `local.db` and `shared.db` databases.
Previously we had a file named `test.db` which stored all of the data
for sioyek. In this commit, we split the database into two files:
* `local.db` contains a table mapping the paths of files to their MD5
checksums
* `shared.db` contains all other data including the last document
locations, marks, bookmarks, highlights and portals. Since we are now
using checksums instead of raw paths to reference documents, this
database file can be shared (as the name suggests) across machines. So
to copy your sioyek data from one machine to another, you just have to
copy `shared.db` from one machine to the other.