Improve desktop file
TryExec allows the system to check if the executable exists.
The %U in Exec is needed for nautilus to pass the file to be opened to
sioyek. Otherwise it's not even show in the "Open with" dialog.
StartupNotify allow different windows of sioyek to be collected under
the same icon in the docker.