Correct argument  in desktop file
Under XDG Desktop Entry specification, %U represent a list of URLs.
sioyek seem (1) not to be able to handle a list of files and (2) not to
be able to understand URL syntax. In fact, when spaces are substituted
with %20, sioyek will not open the file.

This correctly declares the first argument taken by sioyek to be a file
path to a single file.