Add ability to define synchronous javascript commands
Previously we executed javascript commands only asynchronously. Becasue
many of sioyek commands had to be run on the main thread this introduced
a little delay when executing such commands because we had to wait for
the main thread to poll these commands. In this commit we introduce
the ability to define synchronous javascript commands (which is the new
default) and the previous async commands are defined using
"new_async_js_command" option. The new synchronous commands are run on
the main thread and have almost no delay. On the javascript side,
instead of a single "sioyek" object we now have "sioyek" and
"sioyek_async" objects. The user should use the appropriate object based
on whether they defined a synchronous or asynchronous command.