Add option to hide the titlebar on MacOS.
This commit introduces a new option `macos_hide_titlebar` to hide the
titlebar on MacOS.

The window controls (close, minimize) are also hidden but will be show
when the user hover over the top pixels of the window (where the
titlebar would be if shown).

The titlebar is also hidden on the helper window.

In native fullscreen the window controls are shown as normal.