Use the system theme (dark/light mode)
Automatically toggle between dark and light mode depending on the
desired mode from the operating system.

Introduces a new config option `use_system_theme` which if set
automatically set light or dark mode when starting depending on if the
current operating system theme is in light or dark mode. Then, if the
operating system mode changes, the change is reflected in Sioyek.