bugfix - fixes utf8 exception in detect_paper_name
utf8_decode called from detect_paper_name can throw, in particular when
given invalid utf8. Adds a try / catch block in the calling function to
account for this possibility. Behaviour is that invalid utf8 causes the
paper name to be left blank and an error message is logged to stderr.